+++ examples/react-app/src/App.js
@@ -9,17 +9,17 @@ function App() {
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
+        <a className="App-link" href="about.html" rel="noopener noreferrer">
+          About
+        </a>
+        <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
           Learn React
         </a>
       </header>
     </div>
   );
 }
+
+console.log('>> React App');
 
 export default App;