+++ test/cases/js-import-css-same-in-many4/src/app.js
@@ -1,0 +1,3 @@
+import './app.css';
+
+console.log('>> app');