+++ test/cases/js-import-css-same-in-many4/src/home.js
@@ -1,0 +1,3 @@
+import './app';
+
+console.log('>> home');