+++ test/cases/option-css-filename-chunkFilename/src/main.js
@@ -1,0 +1,4 @@
+import './main.css?v=123';
+//import './main.css';
+
+console.log('>> main');