+++ test/cases/option-css-filename-chunkFilename/webpack.config.js
@@ -1,0 +1,36 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+
+      js: {
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+
+      css: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+        chunkFilename: 'assets/css/[name].[contenthash:8].chunk.css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css)/,
+        use: ['css-loader'],
+      },
+    ],
+  },
+};